refactor: restructure project detail layout and introduce ProjectDetailShell
- Moved project detail logic from layout.tsx to a new ProjectDetailShell component for better organization.
- Updated routing and state management to enhance clarity and maintainability.
- Removed unused imports and streamlined the component structure.
